ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
transformers==4.35.2
requests==2.31.0
numpy==1.26.2
facenet-pytorch
